Attempt to make Wellformed compile faster by adding {-# OPTIONS_GHC -O1 #-} in that and other files that compile slow (only when compiling using GHC 8.10.7 Also added Alonzo and Shelley TxOut HasSpec instances.
