+++ libs/cardano-ledger-test/cardano-ledger-test.cabal
@@ -60,6 +60,7 @@ library
         Test.Cardano.Ledger.Constrained.Conway.Epoch
         Test.Cardano.Ledger.Constrained.Conway.Gov
         Test.Cardano.Ledger.Constrained.Conway.GovCert
+        Test.Cardano.Ledger.Constrained.Conway.InstancesBasic
         Test.Cardano.Ledger.Constrained.Conway.Instances
         Test.Cardano.Ledger.Constrained.Conway.Ledger
         Test.Cardano.Ledger.Constrained.Conway.NewEpoch
