+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Conway/LedgerTypes/WellFormed.hs
@@ -1,33 +1,35 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE ExplicitForAll #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE QuantifiedConstraints #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE TypeSynonymInstances #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
+
+#if __GLASGOW_HASKELL__ < 900
+{-# OPTIONS_GHC -O0 #-}
+#endif
 
 module Test.Cardano.Ledger.Constrained.Conway.LedgerTypes.WellFormed where
 
 import Cardano.Ledger.Api
 import Cardano.Ledger.BaseTypes hiding (inject)
 import Cardano.Ledger.CertState
-import Cardano.Ledger.Conway.Rules
+import Cardano.Ledger.Conway.Rules (GovEnv (..))
 import Cardano.Ledger.Credential (Credential)
 import Cardano.Ledger.EpochBoundary (SnapShot (..), SnapShots (..))
 import Cardano.Ledger.Keys (KeyHash, KeyRole (..))
 import Cardano.Ledger.PoolParams (PoolParams (..))
 import Cardano.Ledger.SafeHash ()
-import Cardano.Ledger.Shelley.LedgerState (
-  AccountState (..),
-  EpochState (..),
-  LedgerState (..),
-  NewEpochState (..),
-  UTxOState (..),
- )
+import Cardano.Ledger.Shelley.LedgerState
 import Cardano.Ledger.UTxO (UTxO (..))
 import Constrained hiding (Value)
 import Constrained.Base (hasSize, rangeSize)
@@ -36,14 +38,103 @@ import Test.Cardano.Ledger.Constrained.Conway ()
 import Test.Cardano.Ledger.Constrained.Conway.Instances
 import Test.Cardano.Ledger.Constrained.Conway.LedgerTypes.Specs
 import Test.Cardano.Ledger.Constrained.Conway.PParams (pparamsSpec)
-import Test.Cardano.Ledger.Generic.PrettyCore (PrettyA (prettyA))
-import Test.QuickCheck (Gen, generate)
+import Test.QuickCheck (Gen)
 
 -- ==============================================================
--- The WellFormed class and instances
+-- Generators for all the types found in the Ledger State.
 -- ==============================================================
 
-class (HasSpec ConwayFn t, LedgerEra era ConwayFn) => WellFormed t era where
+ppX :: forall era. EraPP era => Gen (PParams era)
+ppX = genFromSpec @ConwayFn @(PParams era) pparamsSpec
+
+acctX :: Gen AccountState
+acctX = genFromSpec @ConwayFn @AccountState accountStateSpec
+
+psX :: forall era. EraPP era => Gen (PState era)
+psX = do
+  epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
+  genFromSpec @ConwayFn @(PState era) (pstateSpec (lit epoch))
+
+dsX :: forall era. LedgerEra era ConwayFn => Gen (DState era)
+dsX = do
+  acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
+  pools <-
+    genFromSpec @ConwayFn @(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))
+      (hasSize (rangeSize 8 8))
+  genFromSpec @ConwayFn @(DState era) (dstateSpec (lit acct) (lit pools))
+
+vsX :: forall era. EraPP era => Gen (VState era)
+vsX = do
+  epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
+  genFromSpec @ConwayFn @(VState era) (vstateSpec (lit epoch))
+
+csX :: forall era. LedgerEra era ConwayFn => Gen (CertState era)
+csX = do
+  acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
+  epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
+  genFromSpec @ConwayFn @(CertState era) (certStateSpec (lit acct) (lit epoch))
+
+utxoX :: forall era. LedgerEra era ConwayFn => Gen (UTxO era)
+utxoX = do
+  cs <-
+    genFromSpec
+      @ConwayFn
+      @(Map (Credential 'Staking StandardCrypto) (KeyHash 'StakePool StandardCrypto))
+      (hasSize (rangeSize 30 30))
+  genFromSpec @ConwayFn @(UTxO era) (utxoSpec @era (lit cs))
+
+utxostateX :: forall era. LedgerEra era ConwayFn => PParams era -> Gen (UTxOState era)
+utxostateX pp = do
+  certstate <- csX @era
+  genFromSpec @ConwayFn @(UTxOState era) (utxoStateSpec pp (lit certstate))
+
+govenvX :: PParams Conway -> Gen (GovEnv Conway)
+govenvX pp = genFromSpec @ConwayFn @(GovEnv Conway) (govEnvSpec pp)
+
+conwaygovX :: PParams Conway -> Gen (ConwayGovState Conway)
+conwaygovX pp = do
+  env <- genFromSpec @ConwayFn @(GovEnv Conway) (govEnvSpec pp)
+  genFromSpec @ConwayFn @(ConwayGovState Conway) (conwayGovStateSpec pp env)
+
+lsX :: forall era. LedgerEra era ConwayFn => PParams era -> Gen (LedgerState era)
+lsX pp = do
+  acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
+  epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
+  genFromSpec @ConwayFn @(LedgerState era) (ledgerStateSpec pp (lit acct) (lit epoch))
+
+esX :: forall era. LedgerEra era ConwayFn => PParams era -> Gen (EpochState era)
+esX pp = do
+  epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
+  genFromSpec @ConwayFn @(EpochState era) (epochStateSpec pp (lit epoch))
+
+nesX :: forall era. LedgerEra era ConwayFn => PParams era -> Gen (NewEpochState era)
+nesX pp = genFromSpec @ConwayFn @(NewEpochState era) (newEpochStateSpec pp)
+
+snapX :: forall era. Era era => Gen (SnapShot (EraCrypto era))
+snapX = genFromSpec @ConwayFn @(SnapShot (EraCrypto era)) snapShotSpec
+
+snapsX :: forall era. LedgerEra era ConwayFn => PParams era -> Gen (SnapShots (EraCrypto era))
+snapsX pp = do
+  acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
+  epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
+  ls <- genFromSpec @ConwayFn @(LedgerState era) (ledgerStateSpec pp (lit acct) (lit epoch))
+  genFromSpec @ConwayFn @(SnapShots (EraCrypto era)) (snapShotsSpec (lit (getMarkSnapShot ls)))
+
+instanRewX :: forall era. Era era => Gen (InstantaneousRewards (EraCrypto era))
+instanRewX = do
+  acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
+  genFromSpec @ConwayFn @(InstantaneousRewards (EraCrypto era)) (instantaneousRewardsSpec (lit acct))
+
+drepStateX :: forall era. Era era => Gen (DRepState (EraCrypto era))
+drepStateX = do
+  epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
+  genFromSpec @ConwayFn @(DRepState (EraCrypto era)) (drepStateSpec (lit epoch))
+
+-- ==============================================================
+-- The WellFormed class
+-- ==============================================================
+
+class (EraPP era, HasSpec ConwayFn t) => WellFormed t era where
   -- \| Well formed with PParams as input
   wffWithPP :: PParams era -> Gen t
   wffWithPP _ = wff @t @era
@@ -51,113 +142,68 @@ class (HasSpec ConwayFn t, LedgerEra era ConwayFn) => WellFormed t era where
   -- \| Generate a constrained PParams if one is needed
   wff :: Gen t
   wff = do
-    pp <- genFromSpec @ConwayFn @(PParams era) pparamsSpec
+    pp <- ppX @era
     wffWithPP pp
 
-instance LedgerEra era ConwayFn => WellFormed (PParams era) era where
-  wff = genFromSpec @ConwayFn @(PParams era) pparamsSpec
+-- ==============================================================
+-- The WellFormed instances
+-- ==============================================================
+
+instance EraPP era => WellFormed (PParams era) era where
+  wff = ppX @era
   wffWithPP p = pure p
 
-instance LedgerEra era ConwayFn => WellFormed AccountState era where
-  wff = genFromSpec @ConwayFn @AccountState accountStateSpec
+instance EraPP era => WellFormed AccountState era where
+  wff = acctX
+  wffWithPP _ = acctX
 
-instance LedgerEra era ConwayFn => WellFormed (PState era) era where
-  wff = do
-    epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
-    genFromSpec @ConwayFn @(PState era) (pstateSpec (lit epoch))
+instance EraPP era => WellFormed (PState era) era where
+  wff = psX
+  wffWithPP _ = psX
 
-instance LedgerEra era ConwayFn => WellFormed (DState era) era where
-  wff = do
-    acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
-    pools <-
-      genFromSpec @ConwayFn @(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))
-        (hasSize (rangeSize 8 8))
-    genFromSpec @ConwayFn @(DState era) (dstateSpec (lit acct) (lit pools))
+instance (EraPP era, LedgerEra era ConwayFn) => WellFormed (DState era) era where
+  wff = dsX
+  wffWithPP _ = dsX
 
-instance LedgerEra era ConwayFn => WellFormed (VState era) era where
-  wff = do
-    epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
-    genFromSpec @ConwayFn @(VState era) (vstateSpec (lit epoch))
+instance EraPP era => WellFormed (VState era) era where
+  wff = vsX
+  wffWithPP _ = vsX
 
-instance (c ~ EraCrypto era, LedgerEra era ConwayFn) => WellFormed (DRepState c) era where
-  wff = do
-    epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
-    genFromSpec @ConwayFn @(DRepState c) (drepStateSpec (lit epoch))
+instance (EraPP era, LedgerEra era ConwayFn) => WellFormed (CertState era) era where
+  wff = csX
 
-instance LedgerEra era ConwayFn => WellFormed (CertState era) era where
-  wff = do
-    acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
-    epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
-    genFromSpec @ConwayFn @(CertState era) (certStateSpec (lit acct) (lit epoch))
+instance (EraPP era, LedgerEra era ConwayFn) => WellFormed (UTxO era) era where
+  wff = utxoX
 
-instance LedgerEra era ConwayFn => WellFormed (UTxO era) era where
-  wff = do
-    cs <-
-      genFromSpec
-        @ConwayFn
-        @(Map (Credential 'Staking StandardCrypto) (KeyHash 'StakePool StandardCrypto))
-        (hasSize (rangeSize 30 30))
-    genFromSpec @ConwayFn @(UTxO era) (utxoSpec @era (lit cs))
-
-instance LedgerEra era ConwayFn => WellFormed (UTxOState era) era where
-  wffWithPP pp = do
-    certstate <- wff @(CertState era) @era
-    genFromSpec @ConwayFn @(UTxOState era) (utxoStateSpec pp (lit certstate))
+instance (EraPP era, LedgerEra era ConwayFn) => WellFormed (UTxOState era) era where
+  wffWithPP = utxostateX
 
 instance WellFormed (GovEnv Conway) Conway where
-  wffWithPP pp = genFromSpec @ConwayFn @(GovEnv Conway) (govEnvSpec pp)
+  wffWithPP pp = govenvX pp
 
 instance WellFormed (ConwayGovState Conway) Conway where
-  wffWithPP pp = do
-    env <- genFromSpec @ConwayFn @(GovEnv Conway) (govEnvSpec pp)
-    genFromSpec @ConwayFn @(ConwayGovState Conway) (conwayGovStateSpec pp env)
+  wffWithPP pp = conwaygovX pp
 
-instance LedgerEra era ConwayFn => WellFormed (ShelleyGovState era) era where
+instance (EraPP era, LedgerEra era ConwayFn) => WellFormed (ShelleyGovState era) era where
   wffWithPP pp = genFromSpec @ConwayFn @(ShelleyGovState era) (shelleyGovStateSpec pp)
 
-instance LedgerEra era ConwayFn => WellFormed (LedgerState era) era where
-  wffWithPP pp = do
-    acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
-    epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
-    genFromSpec @ConwayFn @(LedgerState era) (ledgerStateSpec pp (lit acct) (lit epoch))
+instance (EraPP era, LedgerEra era ConwayFn) => WellFormed (LedgerState era) era where
+  wffWithPP = lsX
 
-instance (LedgerEra era ConwayFn, c ~ EraCrypto era) => WellFormed (SnapShot c) era where
-  wffWithPP _ = genFromSpec @ConwayFn @(SnapShot (EraCrypto era)) snapShotSpec
-  wff = genFromSpec @ConwayFn @(SnapShot (EraCrypto era)) snapShotSpec
+instance (EraPP era, LedgerEra era ConwayFn) => WellFormed (EpochState era) era where
+  wffWithPP = esX
 
-instance (LedgerEra era ConwayFn, c ~ EraCrypto era) => WellFormed (SnapShots c) era where
-  wffWithPP pp = do
-    acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
-    epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
-    ls <- genFromSpec @ConwayFn @(LedgerState era) (ledgerStateSpec pp (lit acct) (lit epoch))
-    genFromSpec @ConwayFn @(SnapShots (EraCrypto era)) (snapShotsSpec (lit (getMarkSnapShot ls)))
+instance (EraPP era, LedgerEra era ConwayFn) => WellFormed (NewEpochState era) era where
+  wffWithPP = nesX
 
-instance LedgerEra era ConwayFn => WellFormed (EpochState era) era where
-  wffWithPP pp = do
-    epoch <- genFromSpec @ConwayFn @EpochNo epochNoSpec
-    genFromSpec @ConwayFn @(EpochState era) (epochStateSpec pp (lit epoch))
+instance (EraPP era, c ~ EraCrypto era) => WellFormed (SnapShot c) era where
+  wff = snapX @era
 
-instance LedgerEra era ConwayFn => WellFormed (NewEpochState era) era where
-  wffWithPP pp = genFromSpec @ConwayFn @(NewEpochState era) (newEpochStateSpec pp)
+instance (EraPP era, LedgerEra era ConwayFn, c ~ EraCrypto era) => WellFormed (SnapShots c) era where
+  wffWithPP = snapsX @era
 
-instance (LedgerEra era ConwayFn, c ~ EraCrypto era) => WellFormed (InstantaneousRewards c) era where
-  wff = do
-    acct <- genFromSpec @ConwayFn @AccountState accountStateSpec
-    genFromSpec @ConwayFn @(InstantaneousRewards (EraCrypto era)) (instantaneousRewardsSpec (lit acct))
+instance (EraPP era, c ~ EraCrypto era) => WellFormed (InstantaneousRewards c) era where
+  wff = instanRewX @era
 
--- ================================================================
-
-wffIO :: forall t era. (WellFormed t era, PrettyA t) => IO ()
-wffIO = do
-  t <- generate $ wff @t @era
-  putStrLn (show (prettyA t))
-
-wffFam :: forall f era. (WellFormed (f era) era, PrettyA (f era)) => IO ()
-wffFam = do
-  t <- generate $ wff @(f era) @era
-  putStrLn (show (prettyA t))
-
-wffC :: forall f era. (WellFormed (f (EraCrypto era)) era, PrettyA (f (EraCrypto era))) => IO ()
-wffC = do
-  t <- generate $ wff @(f (EraCrypto era)) @era
-  putStrLn (show (prettyA t))
+instance (EraPP era, c ~ EraCrypto era) => WellFormed (DRepState c) era where
+  wff = drepStateX @era
